implemented bounds checking for map

+++ src/com/lvl6/info/Location.java
@@ -2,25 +2,25 @@ package com.lvl6.info;
 
 public class Location {
 
-  private float latitude;
-  private float longitude;
+  private double latitude;
+  private double longitude;
   
   
-  public Location(float latitude, float longitude) {
+  public Location(double latitude, double longitude) {
     this.latitude = latitude;
     this.longitude = longitude;
   }
   
-  public float getLatitude() {
+  public double getLatitude() {
     return latitude;
   }
-  public void setLatitude(float latitude) {
+  public void setLatitude(double latitude) {
     this.latitude = latitude;
   }
-  public float getLongitude() {
+  public double getLongitude() {
     return longitude;
   }
-  public void setLongitude(float longitude) {
+  public void setLongitude(double longitude) {
     this.longitude = longitude;
   }
